Everything But the stationary AI and the countdown timer works holy shit

+++ Assets/Scenes/Chapter1-3/Scripts/Interactables/TriggerCountdown.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 
 public class TriggerCountdown : Interactable
 {
-    private int CountdownTime = 150;
+    private float CountdownTime = 150f;
     public TextMeshProUGUI CountdownUGUI;
     public MonkeyEscape MonkeyEscape;
     public EnemySpawner Chapter1_3EnemySpawner;
